chore: rename router app
